Refactor Framework Component and Icon Bar Integration
- Removed unused imports and commented-out code in framework.tsx.
- Replaced activity bar items with ICON_REGISTRY in IconBar component.
- Updated IconBar to use click handlers for icon interactions.
- Removed MapContentContext and related logic from framework component.
- Simplified tab management and activity handling in framework component.
- Updated tab bar and main editor area rendering logic.
- Removed testData.ts and related activity bar items.
- Updated type definitions for icon bar and tab bar components.
- Changed drag-and-drop library from react-beautiful-dnd to @hello-pangea/dnd.
- Cleaned up unused code and comments across various components.
