+++ src/framework/src/components/iconBar/iconBar.tsx
@@ -1,26 +1,28 @@
-import React from "react"
-import { cn } from "@/utils/utils"
-import { Button } from "@/components/ui/button"
-import { IconBarProps } from "./types"
+import { cn } from '@/utils/utils'
+import { IconBarResourceBinding } from './types'
+import { Button } from '@/components/ui/button'
+import { ICON_REGISTRY } from '../../resource/iconRegistry'
 
-export default function IconBar({ activityBarItems, activeActivity, handleActivityClick }: IconBarProps) {
+export default function IconBar({ currentActiveId, clickHandlers }: IconBarResourceBinding) {
     return (
-        <div className="w-14 bg-slate-900 flex flex-col items-center py-2">
-            {activityBarItems.map((item) => (
+        <div className='w-14 bg-slate-900 flex flex-col items-center py-2'>
+            {ICON_REGISTRY.map(item => (
                 <Button
-                    id={item.id === "user" ? "user-activity-button" : ""}
+                    id={item.id}
                     key={item.id}
-                    variant="ghost"
-                    size="icon"
-                    className={cn(
-                        "w-10 h-10 mb-1 hover:bg-gray-700 cursor-pointer",
-                        activeActivity === item.id && "bg-gray-700 border-l-2 border-blue-500",
-                        item.id === "user" && "bg-cyan-600 border-l-2 border-blue-500 absolute bottom-2 hover:bg-cyan-500",
-                    )}
-                    onClick={() => handleActivityClick(item.id)}
                     title={item.label}
+                    size='icon'
+                    variant='ghost'
+                    onClick={() => clickHandlers[item.id](item.id)}
+                    className={
+                        cn(
+                            'w-10 h-10 mb-1 hover:bg-gray-700 cursor-pointer', // default styles
+                            item.style && item.style,
+                            currentActiveId === item.id && 'bg-gray-700 border-l-2 border-blue-500',
+                        )
+                    }
                 >
-                    <item.icon className="w-8 h-8 text-white"/>
+                    <item.icon className='w-8 h-8 text-white' />
                 </Button>
             ))}
         </div>
