+++ src/framework/src/components/iconBar/types.ts
@@ -1,7 +1,17 @@
-import { ActivityBarItem } from "../framework"
+import { LucideProps } from 'lucide-react'
 
-export interface IconBarProps {
-    activityBarItems: ActivityBarItem[]
-    activeActivity: string
-    handleActivityClick: (id: string) => void
+export interface IconBarClickHandlers {
+    [iconID: string]: (iconID: string) => void
+}
+
+export interface IconBarResourceBinding {
+    currentActiveId: string
+    clickHandlers: IconBarClickHandlers
+}
+
+export interface IconEntry {
+    id: string
+    label: string
+    style?: string
+    icon: React.ComponentType<LucideProps>
 }
