+++ src/framework/src/components/resourceTree/resourceTree.ts
@@ -1,0 +1,5 @@
+// import { SceneNode, ResourceTree } from '@/core/tree/scene'
+
+// class DomSceneNode extends SceneNode {
+//     constructor()
+// }
