+++ src/framework/src/components/types.ts
@@ -1,17 +1,4 @@
-import { LucideProps } from "lucide-react"
-import { ResourceTree } from "@/core/tree/scene";
-
-export interface ActivityBarItem {
-    id: string
-    icon: React.ComponentType<LucideProps>
-    label: string
-    tabName: string
-    name: string
-    path: string
-    isActive: boolean
-    activityId: string
-    isPreview?: boolean
-}
+import { ResourceTree } from '@/core/tree/scene'
 
 export interface Tab {
     id: string
