+++ src/framework/tsconfig.app.json
@@ -27,5 +27,5 @@
     "noFallthroughCasesInSwitch": true,
     "noUncheckedSideEffectImports": true
   },
-  "include": ["src"]
+  "include": ["src", "src/resource/iconRegistry.ts"]
 }
